requirements.txt
streamlit==1.31.0
sentence-transformers==3.0.0
langchain-mistralai==0.0.5
langchain-core==0.1.52
scikit-learn==1.4.0
pypdf==4.0.1
python-docx==1.1.0
python-dotenv==1.0.0
